turn javascript into constant for a single file
