--- packages/shared-assets/best_practices_scripts/progressbar.js
@@ -1,14 +1,0 @@
-const onProgress = (event) => {
-  const progressBar = event.target.querySelector('.progress-bar');
-  const updatingBar = event.target.querySelector('.update-bar');
-  updatingBar.style.width = `${event.detail.totalProgress*100}%`;
-  if (event.detail.totalProgress === 1) {
-    progressBar.classList.add('hide');
-  } else {
-    progressBar.classList.remove('hide');
-    if (event.detail.totalProgress === 0) {
-      event.target.querySelector('.center-pre-prompt')?.classList.add('hide');
-    }
-  }
-};
-document.querySelector('model-viewer').addEventListener('progress', onProgress);
