+++ packages/space-opera/src/components/best_practices/styles.css.ts
@@ -21,15 +21,12 @@ export const modelViewerStyles: CSSResult = css`
 body {
   margin: 1em;
   padding: 0;
-  font-family: Google Sans, Noto, Roboto, Helvetica Neue, sans-serif;
-  color: #244376;
 }
 
 model-viewer {
   width: 100%;
-  height: 400px;
-  background-color: #70BCD1;
-  --poster-color: #ffffff00;
+  height: 100%;
+  background-color: #ffffff;
 }
 `;
